backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.4
langchain-google-genai==0.0.6
google-generativeai==0.3.2
